% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_lidar_files.R
\name{list_lidar_files}
\alias{list_lidar_files}
\title{Provide a list of all lidar files in a given dir}
\usage{
list_lidar_files(dir = "./data/")
}
\arguments{
\item{dir}{file.path}
}
\value{
list of files in dir
}
\description{
Provide a list of all lidar files in a given dir
}
