% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_lidar_rasters.R
\name{merge_lidar_rasters}
\alias{merge_lidar_rasters}
\title{merge all lidar raster}
\usage{
merge_lidar_rasters(lidar_dir = file.path("G:/USGS_lidar"))
}
\arguments{
\item{lidar_dir}{a file.path to a directory containing tiled lidar rasters}
}
\description{
merge all lidar raster
}
