% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lidar_bbox.R
\name{get_lidar_bbox}
\alias{get_lidar_bbox}
\title{This function extract a SpatialPolygon data.frame from the lidar data.frame}
\usage{
get_lidar_bbox(lidar_df)
}
\arguments{
\item{lidar_df}{a data.frame with columns name, networkr, westbc, eastbc, northbc, southbc}
}
\value{
a SpatialPolygon data.frame
}
\description{
This function extract a SpatialPolygon data.frame from the lidar data.frame
}
